Add OpenAISession
