+++ examples/memory/sqlalchemy_session_example.py
@@ -20,28 +20,56 @@ async def main():
         create_tables=True,
     )
 
-    print("=== SQLAlchemySession Example ===")
+    print("=== Session Example ===")
     print("The agent will remember previous messages automatically.\n")
 
     # First turn
+    print("First turn:")
     print("User: What city is the Golden Gate Bridge in?")
     result = await Runner.run(
         agent,
         "What city is the Golden Gate Bridge in?",
         session=session,
     )
-    print(f"Assistant: {result.final_output}\n")
+    print(f"Assistant: {result.final_output}")
+    print()
 
     # Second turn - the agent will remember the previous conversation
+    print("Second turn:")
     print("User: What state is it in?")
+    result = await Runner.run(agent, "What state is it in?", session=session)
+    print(f"Assistant: {result.final_output}")
+    print()
+
+    # Third turn - continuing the conversation
+    print("Third turn:")
+    print("User: What's the population of that state?")
     result = await Runner.run(
         agent,
-        "What state is it in?",
+        "What's the population of that state?",
         session=session,
     )
-    print(f"Assistant: {result.final_output}\n")
+    print(f"Assistant: {result.final_output}")
+    print()
 
     print("=== Conversation Complete ===")
+    print("Notice how the agent remembered the context from previous turns!")
+    print("Sessions automatically handles conversation history.")
+
+    # Demonstrate the limit parameter - get only the latest 2 items
+    print("\n=== Latest Items Demo ===")
+    latest_items = await session.get_items(limit=2)
+    print("Latest 2 items:")
+    for i, msg in enumerate(latest_items, 1):
+        role = msg.get("role", "unknown")
+        content = msg.get("content", "")
+        print(f"  {i}. {role}: {content}")
+
+    print(f"\nFetched {len(latest_items)} out of total conversation history.")
+
+    # Get all items to show the difference
+    all_items = await session.get_items()
+    print(f"Total items in session: {len(all_items)}")
 
 
 if __name__ == "__main__":
