+++ src/agents/memory/__init__.py
@@ -1,3 +1,10 @@
-from .session import Session, SQLiteSession
+from .openai_session import OpenAISession
+from .session import Session, SessionABC
+from .sqlite_session import SQLiteSession
 
-__all__ = ["Session", "SQLiteSession"]
+__all__ = [
+    "Session",
+    "SessionABC",
+    "SQLiteSession",
+    "OpenAISession",
+]
